services/quiz_backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
requests>=2.31.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.9
